Added Ren'Py Nothing
? Quite a process-oriented thing to get this one going including some 
default choices made explicit (color scheme, resolution) before you can 
have your "nothing".

? I exported to web (just feels like the most accessible version for 
people to actually experience) and it was only on playing it that I 
discovered just how much content there is!

? Notable there's a whole UI involved with game loading, preference 
setting. Quite serious stuff.

? And then there's actually a little bit of a "game" when you Start it 
from the menu, with a character called Eileen telling you you need to 
work on making your game. Hah. Also a couple of missing image errors - 
funny there aren't defaults.

? Also to date this is the LARGEST Nothing I have in the project. 
There's a game.zip in the folder which is 2.6MB. Another .gz that's 
3.4MB. And a data file that's 1.2MB. Don't really know what's involved 
in all of there - presumably it's parts of the engine taking up all that 
space? It's possible that game.zip is a distributable version of the 
source code, that someone could open in their own copy of Ren'Py?